glb-redirect: don't pass SKB when looking up listeners
The skb and doff arguments to inet[6]_lookup_listener are used by
SO_REUSEPORT groups to determine which socket to dispatch a packet
to (see [1]). In this context, skb should contain the inner headers
for the lookup to work correctly. We never make the adjustment to
skb, so the selected socket is incorrect (but still from the same
group).

This doesn't matter much to us, since we only care that a valid
socket exists. However, we still have to lug skb around.
Pass NULL for skb, which disables the sk_reuseport behaviour (see [2]).
The result is just as wrong as before, but the code becomes a bit simpler.

1: https://elixir.bootlin.com/linux/v5.0.7/source/net/ipv4/inet_hashtables.c#L279
2: https://elixir.bootlin.com/linux/v5.0.7/source/net/core/sock_reuseport.c#L286